+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [1.2.0-rc1] - unreleased
+## [1.2.0-rc2] - 2024-03-12
 
 ### Added
 
@@ -13,7 +13,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
-- Drop support for Nextcloud <= v27. The differences in the files-API
+- Update to recent @nextcloud/vue
+
+- Drop support for Nextcloud <= v25. The differences in the files-API
   are just too big.
 
 ### Fixed